requirements.txt
langchain[all]
sentence-transformers
streamlit
chromadb
pypdf
python-dotenv